CodeFactor fixes
