+++ osu.Game.Rulesets.Sentakki/SentakkiExtensions.cs
@@ -76,8 +76,8 @@ namespace osu.Game.Rulesets.Sentakki
             point.Y -= origin.Y;
 
             Vector2 ret;
-            ret.X = point.X * MathF.Cos(MathUtils.DegreesToRadians(angle)) + point.Y * MathF.Sin(MathUtils.DegreesToRadians(angle));
-            ret.Y = point.X * -MathF.Sin(MathUtils.DegreesToRadians(angle)) + point.Y * MathF.Cos(MathUtils.DegreesToRadians(angle));
+            ret.X = (point.X * MathF.Cos(MathUtils.DegreesToRadians(angle))) + (point.Y * MathF.Sin(MathUtils.DegreesToRadians(angle)));
+            ret.Y = (point.X * -MathF.Sin(MathUtils.DegreesToRadians(angle))) + (point.Y * MathF.Cos(MathUtils.DegreesToRadians(angle)));
 
             ret.X += origin.X;
             ret.Y += origin.Y;
